+++ Wallet/Sources/Services/Network/Models/OperationApiModel.swift
@@ -10,11 +10,11 @@ public struct OperationApiModel: Codable {
         case income = "INCOME"
         case spending = "SPENDING"
     }
-    public var id: Int?
-    public var walletId: Int?
+    public var id: Int64?
+    public var walletId: Int64?
     public var type: ModelType?
     /** Категория операции */
-    public var categoryId: Int?
+    public var categoryId: Int64?
     public var balance: Decimal?
     public var date: Date?
 }
